refactor: move SelectorResult to tool_dispatch, breaking the module cycle
CodeQL alert 446 (py/unsafe-cyclic-import) flagged the capability <->
tool_dispatch cycle. The TYPE_CHECKING guard made it runtime-safe (false
positive as a runtime hazard), but the cycle existed because SelectorResult
lived in the wrong module: it is the ToolSelector protocol's return type and
depends only on ToolBinding. Relocated next to the protocol; capability.py
re-exports it (explicit `as` form for strict no-implicit-reexport), so all
import sites keep working. Both import orders verified cycle-free.
